requirements.txt
pydantic>=2.5
pandas>=2.0
streamlit>=1.30
rapidfuzz>=3.5
faker>=22.0
ulid-py>=1.1
pytest>=7.4
